requirements.txt
beautifulsoup4==4.12.2
boto3==1.33.11
clip==0.2.0
numpy==1.19.5
openai==1.3.8
pandas==1.1.5
Pillow==10.1.0
requests==2.27.1
SQLAlchemy==1.4.49
streamlit==1.10.0
torch==1.10.2
